% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_bio_sample}
\alias{get_bio_sample}
\title{GET a Bio Sample}
\usage{
get_bio_sample(client, bio_sample_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{bio_sample_id}{The ID of the BioSample}
}
\description{
GET a Bio Sample
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
bio_sample <- ref_client \%>\% get_bio_sample("id")
}

}

