% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_call_set}
\alias{get_call_set}
\title{GET a Call Set}
\usage{
get_call_set(client, call_set_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{call_set_id}{The ID of the Call Set}
}
\description{
GET a Call Set
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
call_set <- ref_client \%>\% get_call_set("id")
}

}

