% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_dataset}
\alias{get_dataset}
\title{GET a Dataset}
\usage{
get_dataset(client, dataset_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{dataset_id}{The ID of the Dataset}
}
\description{
GET a Dataset
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
dataset <- ref_client \%>\% get_dataset("id")
}

}

