% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_expression_level}
\alias{get_expression_level}
\title{GET an Expression Level}
\usage{
get_expression_level(client, expression_level_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{expression_level_id}{The ID of the Expression Level}
}
\description{
GET an Expression Level
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
exp_level <- ref_client \%>\% get_expression_level("id")
}

}

