% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_feature}
\alias{get_feature}
\title{GET a Feature}
\usage{
get_feature(client, feature_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{feature_id}{The ID of the Feature}
}
\description{
GET a Feature
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
feature <- ref_client \%>\% get_feature("id")
}

}

