% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_individual}
\alias{get_individual}
\title{GET an Individual}
\usage{
get_individual(client, individual_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{individual_id}{The ID of the Individual}
}
\description{
GET an Individual
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
individual <- ref_client \%>\% get_individual("id")
}

}

