% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_read_group}
\alias{get_read_group}
\title{GET a Read Group}
\usage{
get_read_group(client, read_group_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{read_group_id}{The ID of the Read Group}
}
\description{
GET a Read Group
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
read_group <- ref_client \%>\% get_read_group("id")
}

}

