% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_read_group_set}
\alias{get_read_group_set}
\title{GET a Read Group Set}
\usage{
get_read_group_set(client, read_group_set_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{read_group_set_id}{The ID of the Read Group Set}
}
\description{
GET a Read Group Set
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
read_group_set <- ref_client \%>\% get_read_group_set("id")
}

}

