% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_reference_id}
\alias{get_reference}
\alias{get_reference_id}
\title{GET a Reference}
\usage{
get_reference(client, reference_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{reference_id}{The ID of the Reference}
}
\description{
GET a Reference
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
reference <- ref_client \%>\% get_reference("id")
}

}

