% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_reference_set}
\alias{get_reference_set}
\title{GET a Reference Set}
\usage{
get_reference_set(client, reference_set_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{reference_set_id}{The ID of the Reference Set}
}
\description{
GET a Reference Set
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
reference_set <- ref_client \%>\% get_reference_set("id")
}

}

