% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_rna_quantification}
\alias{get_rna_quantification}
\title{GET an RNA Quantification}
\usage{
get_rna_quantification(client, rna_quantification_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{rna_quantification_id}{The ID of the RNA Quantification}
}
\description{
GET an RNA Quantification
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
rna_quantification <- ref_client \%>\% get_rna_quantification("id")
}

}

