% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_rna_quantification_set}
\alias{get_rna_quantification_set}
\title{GET an RNA Quantification Set}
\usage{
get_rna_quantification_set(client, rna_quantification_set_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{rna_quantification_set_id}{The ID of the RNA Quantification Set}
}
\description{
GET an RNA Quantification Set
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
rna_quantification_set <- ref_client \%>\% get_rna_quantification_set("id")
}

}

