% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variant}
\alias{get_variant}
\title{GET a Variant}
\usage{
get_variant(client, variant_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{variant_id}{The ID of the Variant}
}
\description{
GET a Variant
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
variant <- ref_client \%>\% get_variant("id")
}

}

