% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_variant_annotation_set}
\alias{get_variant_annotation_set}
\title{GET a Variant Annotation Set}
\usage{
get_variant_annotation_set(client, variant_annotation_set_id)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{variant_annotation_set_id}{The ID of the Variant Annotation Set}
}
\description{
GET a Variant Annotation Set
}
\examples{
## Create a client
ref_client <- ga4gh_client("http://1kgenomes.ga4gh.org", api_location = "")
\dontrun{
library(magrittr)
## Retrieve the object with id 'id'
variant_annotation_set <- ref_client \%>\% get_variant_annotation_set("id")
}

}

