% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_feature_phenotype_associations}
\alias{search_feature_phenotype_associations}
\title{Search for Feature Phenotype Associations}
\usage{
search_feature_phenotype_associations(client, ...)
}
\arguments{
\item{client}{A ga4gh_client object}

\item{...}{Other named arguments for the body of the POST request}
}
\description{
Search for Feature Phenotype Associations
}

