% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ga4gh.R
\name{search_ga4gh}
\alias{search_ga4gh}
\alias{search_ga4gh.default}
\alias{search_ga4gh.ga4ghClient}
\alias{search_ga4gh.list}
\title{Search}
\usage{
search_ga4gh(object, ...)

\method{search_ga4gh}{ga4ghClient}(object, operation, ...)

\method{search_ga4gh}{default}(object, port = NULL, operation, ...)

\method{search_ga4gh}{list}(object, operation, ...)
}
\arguments{
\item{object}{An object to POST to. I.e. A ga4ghClient object, a url for the server or a list of either.}

\item{...}{Named fields for the JSON body}

\item{operation}{The name of an operation. Eg "variantsets"}

\item{port}{The port number}
}
\description{
Search the GA4GH API
}

